etl/requirements.txt
# ============================================================
# ETL Requirements
# ============================================================
pandas>=2.0.0
numpy>=1.24.0
pyodbc>=5.0.0
sqlalchemy>=2.0.0
openpyxl>=3.1.0
xlrd>=2.0.1
python-dateutil>=2.8.0
apscheduler>=3.10.0
python-dotenv>=1.0.0
